refactor(autostart): move platform.py → os_/autostart/
